+++ src/Products/MeetingLiege/profiles/testing/import_data.py
@@ -42,7 +42,6 @@ collegeMeeting.usedItemAttributes = ['budgetInfos',
                                      'toDiscuss',
                                      'completeness',
                                      'otherMeetingConfigsClonableToPrivacy',
-                                     'archivingRef',
                                      'motivation',
                                      'textCheckList',
                                      'itemIsSigned']